DAT-409 - Add Socrata harvester
Adds a harvester for [Socrata Open Data API discovery
endpoints](https://socratadiscovery.docs.apiary.io/#), such as
the one offered by [Open Data Camden](https://opendata.camden.gov.uk).

It can be configured on the new harvester form by:
- Entering a URL - so far I have only tried https://opendata.camden.gov.uk
as that's all we need for this ticket
- Selecting SODA as the source type
- Adding 'app_token' to the configuration, with an application token
  associated with your account on the source site. We have one of these
for production use (see Sven for info), but for test/development it is
free and easy to create one yourself.

Socrata APIs don't have a concept of "dataset" vs "resource" - the
endpoint provides a list of resources, which may have relationships to
other resources. I have ignored these relationships and created one dataset
per resource for now

Files are not downloaded and stored in CKAN, as this would just be a
waste of space and another thing to worry about synchronising - we
just link to the resource on the source.

Not all resources in the Camden source catalogue have a link to an actual
file associated with them. Additionally, some have links that 404 -
another GIGO situation. I have decided to treat both in the same way: if
we can't get a 200 response from the resource download endpoint, we use
the link to the HTML page on the source website where users can view all
available information.

There's no guaruntee that this will work for other Socrata APIs, as not
all of them will have a web interface that works like this, but it works
for the Camden Open Data site and we can cross other bridges when we get to them.

The organisations are given only by name in the source API, so we
generate IDs for them based on the name. This means that we see
different organisations for different spellings of the same thing, such
as "London Borough of Camden Council" and "London Borough of Camden
Counil". This is a garbage-in, garbage-out situation, about which I am
not inclined to do anything. Some source datasets do not have an
associated organisation and will be attributed to whatever organisation
was given as default in the harvester. If the organisations generated
from this system are too messy, we can just omit the 'remote_orgs' key
from the config and use a default organisation for all of them.