DAT-409 - Delete datasets no longer present in source
Check all datasets previously imported by this harvester and compare the IDs to
the list returned from the source so we can delete those that are no longer
present. Since the harvester is now aware of the datasets currently in CKAN, I
have used this information to check which ones are new so we only check for
changes when a dataset already exists and only add the organisation and extras
when it is new.