DAT-409 - Keep extras when updating dataset
When a dataset is updated by the harvester, merging the current and
harvested records overwrites the "extras" field, losing the data quality
and boost factor. This change adds the current values of these fields to
the record before saving the update by using the common util function. I
also refactored the part that adds default values to new dataset records
to use a common util function.