Stop setting modified time to now in datapress harvester
Remove this line, it looks like it was originally added to the
HTMLisation work to force an update but that appears superfluous now
this is handled in reindexing.

This line also likely regressed the earlier behaviour on DAT-340/548
where the previous team overrid CKANs modified time behaviour to mean
when the data was last updated, rather than the metadata.